+++ tools/common/delta/testdata/nonbreaking/new/new.go
