+++ tools/common/go.mod
@@ -1,0 +1,3 @@
+module github.com/Azure/azure-sdk-for-go/tools/common
+
+go 1.13
